finall commit to move tests from P3059R2_* to main-view-test-files;
