+++ tests/std/test.lst
@@ -737,12 +737,6 @@ tests\P2693R1_text_formatting_header_thread
 tests\P2693R1_text_formatting_stacktrace
 tests\P2693R1_text_formatting_thread_id
 tests\P3016R6_inconsistent_begin_end
-tests\P3059R2_join_view_sentinel_constructor
-tests\P3059R2_lazy_split_view_inner_outer_iterators_constructors
-tests\P3059R2_split_view_iterator_sentinel_constructors
-tests\P3059R2_take_view_sentinel_constructor
-tests\P3059R2_take_while_view_sentinel_constructor
-tests\P3059R2_transform_view_iterator_sentinel_constructors
 tests\P3107R5_enabled_specializations
 tests\P3349R1_contiguous_iterators_to_pointers
 tests\P3503R3_packaged_task_promise_with_allocator
