+++ tests/std/tests/P0896R4_views_lazy_split/test.cpp
@@ -3,8 +3,8 @@
 
 #include <algorithm>
 #include <cassert>
-#include <ranges>
 #include <istream>
+#include <ranges>
 #include <span>
 #include <sstream>
 #include <string>
