--- tests/std/tests/P3059R2_take_view_sentinel_constructor/env.lst
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-RUNALL_INCLUDE ..\usual_20_matrix.lst
